modules/dashboard/requirements.txt
streamlit>=1.35.0
snowflake-connector-python>=3.10.0
pandas>=2.2.0
pydeck>=0.9.0
plotly>=5.20.0
python-dotenv>=1.0.0
